requirements.txt
# tensorflow-gpu==1.12
tensorflow==1.13
numpy
scipy
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
tqdm
pyyaml
flake8
jupyter
